Pin 2022-01-07-nightly toolchain
This ensures everyone uses the same toolchain when building to improve
reproducibility of issues. Also, updating to a new toolchain becomes an
intentional change now.

Since cargo takes care of managing the toolchain based on the
rust-toolchain.toml file we don't need the specifics in the bootstrap.sh

Only install rust if `rustc` is not available in bootstrap.sh